+++ addon/components/lt-column-resizer.js
@@ -1,14 +1,11 @@
-import Ember from 'ember';
+import $ from 'jquery';
+import Component from '@ember/component';
+import { computed } from '@ember/object';
 import layout from '../templates/components/lt-column-resizer';
-
-const {
-  $,
-  computed
-} = Ember;
 
 const TOP_LEVEL_CLASS = '.ember-light-table';
 
-export default Ember.Component.extend({
+export default Component.extend({
   layout,
   classNameBindings: [':lt-column-resizer', 'isResizing'],
   column: null,