+++ tests/dummy/app/components/rows/selectable-table.js
@@ -1,12 +1,9 @@
 // BEGIN-SNIPPET selectable-table
-import Ember from 'ember';
+import Component from '@ember/component';
 import TableCommon from '../../mixins/table-common';
+import { computed } from '@ember/object';
 
-const {
-  computed
-} = Ember;
-
-export default Ember.Component.extend(TableCommon, {
+export default Component.extend(TableCommon, {
   hasSelection: computed.notEmpty('table.selectedRows'),
 
   columns: computed(function() {